Added the following functionality:
when creating/deleting/moving an asset, also create/delete/move the thumbnail version of the asset.
delete asset tags for an asset
be able to change to folder not only by name but also by id
modified dimSim.jsp so it can do all DAM functions

Fixed the following bugs:

When deleting a folder/asset, the xml returned still included the deleted folder/asset.
When loosing the ROOT dir or currentFolder == null then xml was creating an exception.