Added the following functionality:
Need to add the ability to create .png thumbnails 

Keep in mind that if an image is png24 make sure the Thumbnail is png24 as well 
and if the image is png8 keep the thumbnail is png8 also.  

Additional thumbnail file GIF type support.

Fixed the following bugs:

when changeToParent after a moveAssetToFolder, folder with the moved asset was not being refreshed from db, making it look like it didn't work.