Added the following functionality:
When a DAM request is missing a parameter or has an invalid parameter name, a more meaningful error msg will come up saying something like the parameters so and so are wrong....or missing. Corrected all command calls.


Fixed the following bugs:

When a DAM request to damXml with no action parameter is sent, return the entire DAM hierarchy structure AND keep the current folder unchanged after the damXml call. Currently, the initial call to damXML returns the entire file structure. However any subsequent calls after an action like changeToFolder or deleteFolder, etc... does not return the entire structure it only returns the hierarchy structure of the current folder. .